+++ common/origin_transforms.cpp
@@ -35,7 +35,7 @@ ORIGIN_TRANSFORMS::~ORIGIN_TRANSFORMS()
 int ORIGIN_TRANSFORMS::ToDisplay( int           aValue,
                                   COORD_TYPES_T aCoordType )
 {
-    return ToDisplay( static_cast<long long int>(aValue), aCoordType );
+    return static_cast<int>( ToDisplay( static_cast<long long int>( aValue ), aCoordType ) );
 }
 
 long long int ORIGIN_TRANSFORMS::ToDisplay( long long int aValue,
@@ -53,7 +53,7 @@ double ORIGIN_TRANSFORMS::ToDisplay( double        aValue,
 int ORIGIN_TRANSFORMS::FromDisplay( int           aValue,
                                     COORD_TYPES_T aCoordType )
 {
-    return FromDisplay( static_cast<long long int>(aValue), aCoordType );
+    return static_cast<int>( FromDisplay( static_cast<long long int>( aValue ), aCoordType ) );
 }
 
 long long int ORIGIN_TRANSFORMS::FromDisplay( long long int aValue,
